requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
mysql-connector-python
python-dotenv
matplotlib
